restructure assetexchange lib; fix contractId API
