+++ core/network/fabric-interop-cc/libs/assetexchange/go.mod
@@ -9,3 +9,5 @@ require (
 	github.com/hyperledger/fabric-protos-go v0.0.0-20210720123151-f0dc3e2a0871
 	github.com/sirupsen/logrus v1.8.1
 )
+
+replace github.com/hyperledger-labs/weaver-dlt-interoperability/common/protos-go => ../../../../../common/protos-go/
